requirements.txt
Django==1.8
PyYAML==3.11
argparse==1.2.1
django-positions==0.5.2
django-rest-swagger==0.2.9
django-uuidfield==0.5.0
djangorestframework==3.1.1
djoser==0.2.1
uuid==1.30
wsgiref==0.1.2
